Add rake tasks to build and install gzip 1.2.7
Gzip is the first build dependency and has no dependencies itself other
than the standard C library.  This patch adds a Rakefile complete with a
`rake build` task that starts the process of building groups of
software.

The end result of a `rake build` with this patch applied is the
following library available for other software to link against:

    $ otool -L /opt/puppet/versions/ruby/ruby-1.9.3-p194/lib/libz.dylib
    /opt/puppet/versions/ruby/ruby-1.9.3-p194/lib/libz.dylib:
        /opt/puppet/versions/ruby/ruby-1.9.3-p194/lib/libz.1.dylib (compatibility version 1.0.0, current version 1.2.7)
        /usr/lib/libSystem.B.dylib (compatibility version 1.0.0, current version 169.3.0)

Additional software build tasks are possible by creating new classes descending
from GenericBuilder.