Add dependency tracking to rake tasks
Without this patch zlib and openssl are compiled and installed every
time the rake build task is run.  This is a problem because the process
is needlessly time consuming.

This patch addresses the problem by adding trivial dependency checking
of high level files installed by the respective tasks. `lib/libz.dylib`
and `bin/openssl`