Add pvm script to activate the shell environment
Without this patch we have Ruby built and installed, but we don't have
an easy way to actually use the software in the new filesystem
hierarchy.

This patch addresses the problem by providing a small script that will
print out sh compatible shell code suitable for an `eval` statement.

For example:

    $ which ruby
    /usr/bin/ruby
    $ eval `/opt/puppet/versions/bin/pvm`
    (1.9.3-p327 dev)$ which ruby
    /opt/puppet/versions/ruby/1.9.3-p327/bin/ruby

Note the prompt is changed as well to indicate the gemset and ruby version
currently activate in the shell environment.