Make pvm work when sub shells are created
Without this patch, sub shells that re-order the PATH will cause PVM to
no longer work properly.  The root cause of this issue is that an
un-intended Ruby may be executed, e.g. `/usr/bin/ruby`, but `GEM_PATH`
and `GEM_HOME` are still set.  This causes compiled Gems such as json or
nokogiri to fail because they are linked against the PVM ruby.

This patch addresses the problem by making sure PVM is last in
~/.zshenv.  This makes PVM incompatible with other MRI rubies like the
system ruby, rvm or rbenv.