Add rake install:gems task
Without this patch there are no default gems installed and as a result
the GEM_HOME and GEM_PATH directories are not being created.  This is a
problem because pvm will need to have bundler in order to install the
rest of the dependencies for projects.  This patch addresses the problem
by making bundler available in the default gemset.

This patch also makes it possible to execute commands directly with the
pvm wrapper script.  This may be desirable in situations where the
parent process environment should not be modified.