Add backline configuration to build /opt/backline
Without this patch we don't have an easy way to build the backline
packages for Mac OS X.  Backline packages are necessary because we need
to connect the Mac builder systems to a
puppetmaster.delivery.puppetlabs.net and classify the system prior to
connecting the builders to jenkins-staging.delivery.puppetlabs.net and
jenkins.delivery.puppetlabs.net.

This patch addresses the problem by adding a new configuration key named
:root_is_prefix that will control the behavior of the GenericBuilder
base class.

This configuration can be built with:

    rake build PVM_CONFIG=backline