Replace unpacked source trees with tarballs
Without this patch we're getting lots of encoding errors with the
resulting compiled rubies.  This is a problem because the errors cause
unhandled exceptions and we need all of the encodings to properly
support string conversions.

This patch addresses the problem by converting the large number of
unpacked files in the git repository into a single tarball.  This is
mentioned as a potential solution to this problem at
http://www.pressingquestion.com/4210120/What-Causes-A-Encodingconverter