Build each config into a separate package
Without this patch it's difficult to compose multiple runtime versions
into a complete toolchain with multiple ruby versions.  This patch
addresses the problem by adding a name for each configuration.  The
package built will now be the name of the configuration with an appended
version identifier based on `git describe`.