Add purge:root and purge:destroot tasks
Without this patch it's difficult to clean up after each build and
package cycle for a particular configuration.  A simple task to remove
both the system root location and the destroot location partially clean
up after each build.  What remains is the source in src/, but a `git
clean -fdx src`, `git reset --hard` fixes this.