Add Mac OS X version to package version
Without this patch the version number includes the Mac OS X OS version
at the end of the package name.  This is a problem because the OS
version is the most significant piece of information when selecting a
version to install and having it as a suffix may cause more recent
versions on an earlier OS to be installed.  This patch addresses the
problem by moving the OS version to the front which will prevent
packages from an earlier OS from being installed on the latest OS.