Add rake synthesize task to speed up builds
Without this patch it's difficult to synthesize a bunch of individual
builds because the behavior is at the end of the large rake crossfader
task.  This patch addresses the problem by moving the build and
synthesize behaviors into their own builder class instance method.  A
new task calls the instance method on demand.