Fix side effect on package_id method
Without this patch calling the package_id method has the side effect of
changing the name.  This is a problem because the wrong name is used to
produce the package after the package_id method is called.

The fix is to eliminate the side effect by creating a new string
instance.