Write final package artifact to artifacts/
Without this patch the build archives are growing quite large.  We're
copying a lot of data we don't really need to archive after each build.
The only thing that really matters is the distribution archive itself so
this patch changes the crossfader task to write the distribution package
to the artifacts/*.pkg directory.