Add xfade-run script to setup environment
Without this patch there isn't an easy way to install Gem dependencies
into the crossfader runtime environment itself.  This is a problem
because we need a way to easily run commands inside of an environment
that has all of the libraries we need.

We can't simply install to the system gem location since we need to keep
that gem path clean for other uses.

This patch addresses the problem by initializing the gem home at
/opt/crossfader/runtime/gemsets/crossfader.  In addition the global
gemset is initialized.

This should help packaging tasks that can produce an updated tools
package by simply packaging up the
/opt/crossfader/runtime/gemsets/crossfader directory.