Fix rake task unable to execute commands
Without this patch we're seeing failures we weren't previously seeing in
the Jenkins builds.  The failures look like:

    $ bundle exec rake crossfader --trace
    ** Invoke crossfader (first_time)
    ** Execute crossfader
    rm -rf pkg
    rm -rf destroot
    rm -rf artifacts
    git clean -fdx src/
    git checkout HEAD src/
    rm -rf /opt/crossfader/runtime
    bundle exec rake PVM_CONFIG=crossfaderuntime build
    rake aborted!
    Command failed with status (127): [bundle exec rake PVM_CONFIG=crossfaderunti...]
    /Users/jenkins/workspace/Crossfader-Package-MacOSX-10.8/.bundle/gems/ruby/1.9.1/gems/rake-10.1.0/lib/rake/file_utils.rb:55:in `block in create_shell_runner'
    /Users/jenkins/workspace/Crossfader-Package-MacOSX-10.8/.bundle/gems/ruby/1.9.1/gems/rake-10.1.0/lib/rake/file_utils.rb:45:in `call'
    /Users/jenkins/workspace/Crossfader-Package-MacOSX-10.8/.bundle/gems/ruby/1.9.1/gems/rake-10.1.0/lib/rake/file_utils.rb:45:in `sh'
    /Users/jenkins/workspace/Crossfader-Package-MacOSX-10.8/.bundle/gems/ruby/1.9.1/gems/rake-10.1.0/lib/rake/file_utils_ext.rb:37:in `sh'

I believe this is caused by a change in behavior where by the shell command
being executed is passed as a single string and as such the environment
variables for GEM_HOME and GEM_PATH are not being exported correctly.  This
patch addresses the problem by passing the command to execute as an array,
thereby avoiding the shell since the sh command follows the same semantics as
Kernel.exec()