Fix backline package name
Without this patch the existing backline packages are missing encoding
translators which is a problem because it's interfering with pry's
ability to display interactive prompts.  This in turn is making it
really difficult to build crossfader since there's a problem I'm having
difficult debugging because pry doesn't work right.

This patch addresses the problem by updating the backline configuration
to use the new build names.  This should result in a backline package
with bundler installed and all of the encodings properly generated.