Stop using bundler because it interferes with gem installation
Without this patch it's difficult to use the parent rake task to build
all of the packages because Bundler starts off and pollutes the
environment.  This environment pollution prevents the gem install
command from installing the bundler gem into the correct location.
Instead it installs into the bundle directory for the build.

This patch fixes the problem by removing all situations where bundler is
used during the build process.  The CI job is expected to use rake
directly.  Bundler may still be used to install dependencies, but the
job is expected to do something like this:

    #!/bin/bash
    set -x
    set -e

    sudo rm -rf /opt/crossfader
    sudo mkdir /opt/crossfader
    sudo chown $USER:0 /opt/crossfader
    sudo chmod 755 /opt/crossfader

    # NOTE: This job specifically avoids using bundler since it interferes with the rake
    # tasks that installs gems into the target directory that is meant to be packaged.
    eval "$(/opt/backline/bin/backline shellinit)"
    bundle install --path .bundle/gems/
    bundle clean

    # Don't use bundle exec, it pollutes the shell environment
    export GEM_PATH="$(pwd)/.bundle/gems/ruby/1.9.1:${GEM_PATH}"
    export PATH="$(pwd)/.bundle/gems/ruby/1.9.1/bin:{$PATH}"

    echo "Building crossfader packages..."
    rake crossfader