Refactor crossfader into a full gem
Without this patch it's difficult to install the crossfader command line
utility without doing so with a hacked together rake task or what have
you.  This is a problem because it needs to be relatively easy to update
the user facing command line utility itself.

This patch addresses the problem by transforming the crossfader command
into a full fledged Gem.  Installation of the utility into the runtime
environment is now a simple `gem install` and everything goes into the
right place with the right dependencies.

The only additional task required is to link
/opt/crossfader/bin/crossfader to the installed gem executable at
/opt/crossfader/runtime/bin/crossfader.  This is necessary to keep the
runtime ruby environment out of the general PATH but still easily
accessible from a self contained prefix installation location.

It's important to note this design and layout only apply to the Ruby
runtime for crossfader itself.  The ruby versions that are activated by
the crossfader tool are still independent and will have unpolluted
system gem repositories.