(maint) Fix crossfader exception after shellinit
Without this patch crossfader works once, before the shell has been
initialized to a specific ruby version but it fails to run after shell
initialization.  The following error hits the user:

    $ crossfader --help
    /opt/crossfader/runtime/lib/ruby/1.9.1/rubygems/dependency.rb:247:in `to_specs': Could not find crossfader (>= 0) amongst [bigdecimal-1.1.0, io-console-0.3, json-1.5.5, minitest-2.5.1, rake-0.9.2.2, rdoc-3.9.5] (Gem::LoadError)
      from /opt/crossfader/runtime/lib/ruby/1.9.1/rubygems/dependency.rb:256:in `to_spec'
      from /opt/crossfader/runtime/lib/ruby/1.9.1/rubygems.rb:1231:in `gem'
      from /opt/crossfader/runtime/bin/crossfader:22:in `<main>'

The root cause of this problem is that crossfader is installed into the default
GEM location for the Ruby in `/opt/crossfader/runtime/bin/crossfader` and the
GEM_HOME and GEM_PATH set by running `crossfader shellinit` interfere with
crossfader's ability to load itself.

This patch addresses the problem by clearing the three environment variables
prior to running the crossfader command.  This ensures the "runtime" Ruby is
used in it's default state and no other versions are interfering with the
system.