Make ruby gemset directories admin writable
Without this patch end users of crossfader cannot install gems without
dropping to root access levels.  This is a problem because sudo resets
the environment which makes usability more difficult.

This patch fixes the problem by packaging crossfader in a manner that
uses recommended permissions for nearly everything except the gemset
directory which is root:admin with 775 unix permissions.