Add version identifier to /opt/crossfader/version.txt
Without this patch there is no version identifier for the installed
version of the full crossfader package.  This is a problem because the
crossfader executable command returns the gem version of crossfader, not
the actual build version.  We need to store and write the build version
in the package in order to display it to the user.

This patch addresses the problem by writing
/opt/crossfader/version.txt when building and before packaging
the crossfader runtime environment itself.  This file is managed by a
small, special package with the same version as the single file written.