Display build version from `crossfader --version`
Without this patch the end user doesn't have an way to easily check
which version of crossfader is currently installed.  When we say
"crossfader version" we generally mean the build version, which is all
of the components in the main distribution package.  This patch augments
the crossfader gem version with an additional build version read from
/opt/crossfader/version.txt