Set CROSSFADER_RUBY during shellinit
Without this patch the CROSSFADER_RUBY environment variable is not set
when initializing the shell environment for crossfader.  This makes it
difficult to hand off the currently initialized environment to a sudo
initialized environment with something like `sudo crossfader
--ruby=$CROSSFADER_RUBY --gemset=$CROSSFADER_GEMSET`.

This patch addresses the problem by setting CROSSFADER_RUBY just like
CROSSFADER_GEMSET is defined during shell initialization.

In addition this patch uses XFADE_PATH_ORIG if it's present in the
enviornment.  This has the behavior of using the original PATH when
constructing a new PATH from a nested invocation of crossfader.