Add Ruby build for /opt/operations/bin/ruby
Without this patch the Operations team doesn't have a Ruby interpreter
to use for the purpose of running Puppet to manage Mac OS X
workstations.  This is a problem because the Ops team wants to help
improve the overall experience of supporting other employees in their
efforts.  As part of this improvement they'd like to use Puppet to
automate the configuration of laptops.

This patch addresses the problem by providing a dedicated Ruby as part
of crossfader.  The Ops team will have /opt/operations/bin/ruby for
their own needs.

This build requires the following setup phase in the Jenkins job to
build correctly.  This behavior should probably be moved to a `rake
crossfader:setup` task.

    sudo rm -rf /opt/operations
    sudo mkdir /opt/operations
    sudo chown $USER:0 /opt/operations
    sudo chmod 755 /opt/operations

Paired With: Matt Kirby <kirby@puppetlabs.com>