Fix missing crossfader ruby-1.9.3-p484 version
Without this patch there is no 1.9.3-p484 version packaged for
crossfader.  This is a problem because we need this version to validate
against Puppet Enterprise.

The reason this version disappeared is because the ruby version for the
operations team in /opt/operations shares the same package identifier as
the version for crossfader.

This patch addresses the problem by creating a unique package identifier
string for the operations version.