Update README with package installation guide
Without this patch it's not clear how to get started with crossfader
from an end-user perspective.  The README caters to building the
packages, which is a problem because it's not the fastest path to
success.

This patch addresses the problem by adding a quick start guide for new
users rather than package maintainers.