Fix Rubygems with Ruby 1.8.7
Without this patch rubygems is not properly installed for the Ruby 1.8.7
version included with crossfader.  This is a problem because crossfader
cannot easily bootstrap application dependencies using bundler when
using Ruby 1.8.7.  This is often necessary for QA/QE work with beaker.

This patch addresses the problem by modifying the Ruby builder class to
check if there is a :rubygems section of the given configuration file.
If there is, then rubygems is built before the default set of gems is
installed.

This section should be omitted for all ruby versions after 1.8.7 as
Rubygems is included with MRI itself.